mlp - RHEL compilers don\'t get unint16_t defs... changed

+++ newbasic/oncsSub_idinttv0.h
@@ -47,7 +47,7 @@ protected:
 
   struct intt_hit
   {
-    uint64_t bco;
+    unsigned long long bco;
     unsigned short fee;
     unsigned short channel_id;
     unsigned short chip_id;
